fix(test/tap): real bugs from static analysis
- fix uninitialized loop var (UB) in test_digest_umap_aux-t.cpp
- fix dead stores and unused vars in test_binlog_reader-t.cpp
- rename reserved identifiers (__timer, _mysql_hdr, th_args__in_t, etc.)
- fix narrowing conversions (ssize_t->int, uint32_t->int, size_t->int)
- fix implicit widening of multiplication (overflow risk)
- consolidate duplicated conditional branch in reg_test_4556
- pass modname by const ref in test_cluster1-t.cpp
- fix narrowing in test_default_conn_collation-t.cpp

+++ test/tap/tests/test_default_conn_collation-t.cpp
@@ -43,7 +43,7 @@ int check_all_collations(const CommandLine& cl, MYSQL* admin) {
 		}
 	}
 
-	plan(num_tests);
+	plan(static_cast<int>(num_tests));
 
 	do {
 		if (c[0].nr > 255) {
